Create Feature Branch for User Story 01
As the **Open To-do API**
I want to **return** JSON representations of users, lists, and items
In order to create a framework for working with task lists from the
command line and browser

This commit creates a directory for project documentation (`docs`) and
moves the `PROJECT.md` file to the `docs` directory. It adds a new
document describing planned steps to implement the user story
(`USERSTORY01.md`).